dataset_creation/requirements.txt
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
decorator==4.4.2
future==0.18.2
google-images-download==2.8.0
h5py==2.10.0
joblib==0.16.0
kiwisolver==1.2.0
matplotlib==3.3.1
networkx==2.5
nltk==3.5
numpy==1.19.1
Pillow==7.2.0
pkg-resources==0.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
regex==2020.7.14
selenium==3.141.0
six==1.15.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.48.2
urllib3==1.25.10
